pyproject.toml
[tool.poetry]
name = "lorikeet"
version = "0.1.7"
# Don't forget to also change the version in:
# - package.json
# - docs/conf.py
description = "A simple, generic, API-only shopping cart framework for Django"
authors = ["Adam Brenecki <adam@brenecki.id.au>"]
license = "MIT"
readme = "README.rst"
homepage = "https://lorikeet.readthedocs.io/"
repository = "https://gitlab.com/abre/lorikeet"

[tool.poetry.dependencies]
python = ">=3.4"
django-model-utils = "^2.6"
djangorestframework = "~3.5.3"
stripe = {version = "^1.44.0", optional = true}
premailer = {version = "^3.0.1", optional = true}
requests = {version = "^2.13.0", optional = true}

[tool.poetry.dev-dependencies]
sphinx = "~1.5.1"
sphinx-rtd-theme = "^0.1.9"
sphinx-js = "~1.3"
sphinxcontrib_httpdomain = "^1.5.0"
factory-boy = "^2.11"
tox = "^3.1"

[tool.poetry.extras]
stripe = ["stripe"]
email_invoice = ["premailer"]
starshipit = ["requests"]
